Cast error to integer for comparisons. Closes #235

+++ Model/Behavior/UploadBehavior.php
@@ -552,7 +552,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -577,7 +577,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -602,7 +602,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -628,7 +628,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -675,7 +675,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -700,7 +700,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -726,7 +726,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -761,7 +761,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -796,7 +796,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -848,7 +848,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -884,7 +884,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -920,7 +920,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
@@ -956,7 +956,7 @@ class UploadBehavior extends ModelBehavior {
 			return true;
 		}
 
-		$error = Hash::get($check[$field], 'error');
+		$error = (int) Hash::get($check[$field], 'error');
 
 		// Allow circumvention of this rule if uploads is not required
 		if (!$requireUpload && $error === UPLOAD_ERR_NO_FILE) {
